test.sh: don’t overwrite test.txt/test.compressed when builddir == srcdir

+++ test/test.sh
@@ -7,7 +7,9 @@ rm -f test.txt_ test.new
 
 # We need to copy the source files because the output file will be placed in
 # the same directory. However, $srcdir might not be writable, while $PWD is.
-cp $srcdir/test.txt $srcdir/test.compressed .
+if [ ! -e "test.txt" ]; then
+    cp $srcdir/test.txt $srcdir/test.compressed .
+fi
 
 $mscompress_abspath/src/mscompress test.txt
 $mscompress_abspath/src/msexpand < test.txt_ > test.new
